Split off more deprecations
