+++ codebase2/codebase-sqlite/unison-codebase-sqlite.cabal
@@ -1,6 +1,6 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.36.0.
+-- This file has been generated from package.yaml by hpack version 0.37.0.
 --
 -- see: https://github.com/sol/hpack
 
@@ -52,6 +52,8 @@ library
       U.Codebase.Sqlite.LocalizeObject
       U.Codebase.Sqlite.NamedRef
       U.Codebase.Sqlite.NameLookups
+      U.Codebase.Sqlite.NamesPerspectives.Operations
+      U.Codebase.Sqlite.NamesPerspectives.Queries
       U.Codebase.Sqlite.ObjectType
       U.Codebase.Sqlite.Operations
       U.Codebase.Sqlite.Orphans
