+++ parser-typechecker/src/Unison/Codebase.hs
@@ -71,9 +71,6 @@ module Unison.Codebase
     getBranchPartialDeclNameLookup,
     getBranchDeclNameLookup,
 
-    -- * Root branch
-    SqliteCodebase.Operations.namesAtPath,
-
     -- * Patches
     SqliteCodebase.Operations.patchExists,
     SqliteCodebase.Operations.getPatch,
